% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatCon.R
\name{MatCon}
\alias{MatCon}
\title{Confusion matrix}
\value{
Object of class MatCon #y mas
}
\description{
Using the confusion matrix, various indices are calculated.
}
\examples{
C = matrix( c(5, 0, 1, 0,4,0,0,0,3), nrow=3,  ncol=3)
mc2 <- MatCon$new (C,ID=5,Date="27-10-2023")

## ------------------------------------------------
## Method `MatCon$oa`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$oa()

## ------------------------------------------------
## Method `MatCon$dec`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$dec()

## ------------------------------------------------
## Method `MatCon$ua`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A,ID=1,Date="30/10/2023")
p$ua()

## ------------------------------------------------
## Method `MatCon$uai`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$uai(2)

## ------------------------------------------------
## Method `MatCon$pai`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$pai(1)

## ------------------------------------------------
## Method `MatCon$pa`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$pa()

## ------------------------------------------------
## Method `MatCon$aup`
## ------------------------------------------------

A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$aup(2)
}
\references{
Story, M., & Congalton, R. G. (1986). Accuracy assessment: a user’s perspective. Photogrammetric Engineering and remote sensing, 52(3), 397-399.

Story, M., & Congalton, R. G. (1986). Accuracy assessment: a user’s perspective. Photogrammetric Engineering and remote sensing, 52(3), 397-399.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MatCon-new}{\code{MatCon$new()}}
\item \href{#method-MatCon-oa}{\code{MatCon$oa()}}
\item \href{#method-MatCon-dec}{\code{MatCon$dec()}}
\item \href{#method-MatCon-ua}{\code{MatCon$ua()}}
\item \href{#method-MatCon-uai}{\code{MatCon$uai()}}
\item \href{#method-MatCon-pai}{\code{MatCon$pai()}}
\item \href{#method-MatCon-pa}{\code{MatCon$pa()}}
\item \href{#method-MatCon-aup}{\code{MatCon$aup()}}
\item \href{#method-MatCon-clone}{\code{MatCon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-new"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$new(values, ID = NULL, Date = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{Confusion matrix}

\item{\code{ID}}{Identifier (optional)}

\item{\code{Date}}{System or user-provided date (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-oa"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-oa}{}}}
\subsection{Method \code{oa()}}{
Overall accuracy for a particular classified image/map is then calculated by dividing the sum of the entries that form the major diagonal (i.e., the number of correct classifications) by the total number of samples taken.


The mathematical expression is:

\deqn{
oa = \frac{\sum_{i=1}^{n} x_{ii}}{\sum_{i, j=1}^{n} x_{ij}}
}

Where:
\enumerate{
  \item `oa`: overall accuracy.
  \item `x_ii`: diagonal element of the matrix.
  \item `x_ij`: element of the matrix.
}
This represents a mathematical expression with a fraction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$oa(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$oa()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-dec"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-dec}{}}}
\subsection{Method \code{dec()}}{
Determines whether a value is decimal or not.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$dec(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE or FALSE
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$dec()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-ua"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-ua}{}}}
\subsection{Method \code{ua()}}{
The accuracy from the point of view of a map user, not the map maker.


The mathematical expression is:
\deqn{
ua=\frac{x_{ii}}{\sum_{j=1}^n x_{ij}}
}

where:

\enumerate{
  \item `ua`: user accuracy.
  \item `x_ii`: diagonal element of the matrix.
  \item `x_ij`: element of the matrix.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$ua(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of values with the user's accuracy indexes of all classes
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A,ID=1,Date="30/10/2023")
p$ua()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-uai"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-uai}{}}}
\subsection{Method \code{uai()}}{
The accuracy from the point of view of a map user, not the map maker.


\deqn{
ua_{i}=\frac{x_{ii}}{\sum_{j=1}^n x_{ij}}
}

where:

\enumerate{
  \item `ua_i`: user accuracy.
  \item `x_ii`: diagonal element of the matrix.
  \item `x_ij`: element of the matrix.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$uai(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{User class to evaluate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Class i user accuracy index
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$uai(2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-pai"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-pai}{}}}
\subsection{Method \code{pai()}}{
The map accuracy from the point of view of the map maker (the producer).


\deqn{
pa_{i}=\frac{x_{jj}}{\sum_{j=1}^n x_{ij}}
}

where:

\enumerate{
  \item `pa_i`: producer accuracy.
  \item `x_jj`: diagonal element of the matrix.
  \item `x_ij`: element of the matrix.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$pai(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Producer class to evaluate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Class i producer accuracy index
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$pai(1)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-pa"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-pa}{}}}
\subsection{Method \code{pa()}}{
The map accuracy from the point of view of the map maker (the producer).


\deqn{
pa=\frac{x_{jj}}{\sum_{j=1}^n x_{ij}}
}

where:

\enumerate{
  \item `pa`: producer accuracy.
  \item `x_jj`: diagonal element of the matrix.
  \item `x_ij`: element of the matrix.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$pa(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of values with the producer's accuracy indexes of all classes
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$pa()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-aup"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-aup}{}}}
\subsection{Method \code{aup()}}{
Average of the accuracy from the point of view of a map user, not the map maker and the map accuracy from the point of view of the map maker (the producer).


\deqn{
aup=\frac{uai+pai}{2}
}

where:

\enumerate{
  \item `aup`: Average of user's and producer's accuracy.
  \item `uai`: user accuracy
  \item `pai`: producer accuracy.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$aup(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Class to evaluate.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(36,1,0,0,2,0,0,1,20),nrow=3,ncol=3)
p<-MatCon$new(A)
p$aup(2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatCon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MatCon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatCon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
